+++ packages/plugins/Debugger/src/Worker/index.ts
@@ -1,0 +1,9 @@
+import type { Plugin } from '@masknet/plugin-infra'
+import { base } from '../base'
+import '../messages'
+
+const worker: Plugin.Worker.Definition = {
+    ...base,
+    init(signal) {},
+}
+export default worker
