+++ packages/plugins/Debugger/src/index.ts
@@ -10,4 +10,9 @@ registerPlugin({
         hotModuleReload: (hot) =>
             import.meta.webpackHot && import.meta.webpackHot.accept('./SNSAdaptor', () => hot(import('./SNSAdaptor'))),
     },
+    Worker: {
+        load: () => import('./Worker'),
+        hotModuleReload: (hot) =>
+            import.meta.webpackHot && import.meta.webpackHot.accept('./Worker', () => hot(import('./Worker'))),
+    },
 })
