+++ packages/shared/src/contexts/common/index.tsx
@@ -1,14 +1,18 @@
 import type { FC, PropsWithChildren } from 'react'
 import { ConfirmProvider } from './Confirm'
 import { SelectFungibleTokenProvider } from './SelectFungibleTokenDialog'
+import { SelectGasSettingsProvider } from './SelectGasSettingsDialog'
 
 export * from './Confirm'
 export * from './SelectFungibleTokenDialog'
+export * from './SelectGasSettingsDialog'
 
 export const CommonUIProvider: FC<PropsWithChildren<{}>> = ({ children }) => {
     return (
         <ConfirmProvider>
-            <SelectFungibleTokenProvider>{children}</SelectFungibleTokenProvider>
+            <SelectFungibleTokenProvider>
+                <SelectGasSettingsProvider>{children}</SelectGasSettingsProvider>
+            </SelectFungibleTokenProvider>
         </ConfirmProvider>
     )
 }
