+++ packages/shared/src/contexts/components/index.ts
@@ -1,1 +1,3 @@
 export * from './InjectedDialog'
+export * from './SelectFungibleTokenDialog'
+export * from './SelectGasSettingsDialog'
